Refactor engine logic into separate modules
refs: https://github.com/frostburn/panel-league/issues/43
