+++ lib/engine-basic.js
@@ -1,0 +1,321 @@
+const JKISS31 = require('./jkiss');
+const {newBlock, clearBlock, floodFill} = require('./engine-util');
+
+function blocksCanSwap(block1, block2) {
+  if (!block1 || !block2) {
+    return false;
+  }
+  if (!block1.color && !block2.color) {
+    return false;
+  }
+  if (block1.flashTimer >= 0 || block2.flashTimer >= 0) {
+    return false;
+  }
+  if (block1.floatTimer > 0 || block2.floatTimer > 0) {
+    return false;
+  }
+  if (block1.garbage || block2.garbage) {
+    return false;
+  }
+  return true;
+}
+
+function blocksMatch(block1, block2) {
+  if (!block1 || !block2) {
+    return false;
+  }
+  if (block1.flashTimer >= 0 || block2.flashTimer >= 0) {
+    return false;
+  }
+  if (block1.floatTimer >= 0 || block2.floatTimer >= 0) {
+    return false;
+  }
+  if (!block1.color || !block2.color) {
+    return false;
+  }
+  if (block1.swapTimer !== 0 || block2.swapTimer !== 0) {
+    return false;
+  }
+  if (block1.garbage || block2.garbage) {
+    return false;
+  }
+  if (block1.color === block2.color) {
+    return (!block1.preventMatching && !block2.preventMatching);
+  }
+  return false;
+}
+
+function handleSwapping(state) {
+  state.blocks.forEach((block) => {
+    if (block.swapTimer > 0) {
+      --block.swapTimer;
+    } else if (block.swapTimer < 0) {
+      ++block.swapTimer;
+    }
+  });
+}
+
+// Match three or more similar blocks horizontally or vertically
+function findMatches(state, blocks) {
+  let matchFound = false;
+  blocks.forEach((block, i) => {
+    const bellow = blocks[i + state.width];
+    const above = blocks[i - state.width];
+    let left;
+    let right;
+
+    if (i % state.width > 0) {
+      left = blocks[i - 1];
+    }
+    if (i % state.width < state.width - 1) {
+      right = blocks[i + 1];
+    }
+
+    if (blocksMatch(left, block) && blocksMatch(block, right)) {
+      left.matching = true;
+      block.matching = true;
+      right.matching = true;
+      matchFound = true;
+    }
+
+    if (blocksMatch(bellow, block) && blocksMatch(block, above)) {
+      above.matching = true;
+      block.matching = true;
+      bellow.matching = true;
+      matchFound = true;
+    }
+  });
+  return matchFound;
+}
+
+function invalidateMatches(blocks) {
+  blocks.forEach((block) => {
+    if (block.matching) {
+      block.preventMatching = true;
+    }
+  });
+}
+
+function clearMatches(blocks, includeInvalid) {
+  blocks.forEach((block) => {
+    delete block.matching;
+    if (includeInvalid) {
+      delete block.preventMatching;
+    }
+  });
+}
+
+function pushRow(blocks, nextRow, width) {
+  for (let i = 0; i < width; ++i) {
+    blocks.shift();
+    blocks.push(nextRow[i]);
+  }
+}
+
+// When adding rows we must not create new matches unless forced to.
+function addRow(state) {
+  if (state.nextRow) {
+    pushRow(state.blocks, state.nextRow, state.width);
+    // Find matches that are forced and exclude them.
+    findMatches(state, state.blocks);
+    invalidateMatches(state.blocks);
+    clearMatches(state.blocks);
+  }
+  const RNG = JKISS31.unserialize(state.RNG);
+  while (true) {
+    const nextRow = [];
+    for (let i = 0; i < state.width; ++i) {
+      const block = newBlock();
+      block.color = state.blockTypes[RNG.step() % state.blockTypes.length];
+      nextRow.push(block);
+    }
+    // Make sure that no unnecessary matches would be made when pushing the new row.
+    const candidateBlocks = state.blocks.slice();
+    pushRow(candidateBlocks, nextRow, state.width);
+    const candidateMatches = findMatches(state, candidateBlocks);
+    clearMatches(candidateBlocks);
+    if (!candidateMatches) {
+      state.nextRow = nextRow;
+      break;
+    }
+  }
+  clearMatches(state.blocks, true);
+  state.RNG = RNG.serialize();
+}
+
+function handleEvents(state, events) {
+  const garbage = state.garbage;
+  const blocks = state.blocks;
+
+  events.slice().sort().forEach((event) => {
+    const [, type, param] = event;
+
+    if (type === 'swap') {
+      if (param % state.width < state.width - 1) {
+        const block1 = blocks[param];
+        const block2 = blocks[param + 1];
+
+        if (blocksCanSwap(block1, block2)) {
+          // Upon swapping the blocks immediately warp into their new locations,
+          // but receive a swap timer that partially disable them for a while.
+          // The UI will display this time as a sliding animation.
+
+          // Block 1 goes left to right
+          block1.swapTimer = state.swapTime;
+          // Block 2 goes right to left
+          block2.swapTimer = -state.swapTime;
+
+          blocks[param] = block2;
+          blocks[param + 1] = block1;
+        }
+      }
+    } else if (type === 'addRow') {
+      if (blocks.slice(0, state.width).every(block => !block.color)) {
+        addRow(state);
+        garbage.forEach(slab => ++slab.y);
+      }
+    } else if (type === 'addGarbage') {
+      const slab = param;
+      if (slab.x + slab.width > state.width) {
+        throw new Error('Invalid garbage slab');
+      }
+      // Place the slab on top of all the other slabs or at the top of the screen.
+      slab.y = garbage.reduce(
+        (max, s) => Math.max(max, s.y + s.height),
+        state.height
+      );
+      slab.flashTime = state.garbageFlashTime * slab.width * slab.height;
+      slab.flashTimer = -1;
+      garbage.push(slab);
+    }
+  });
+}
+
+function handleGravity(state) {
+  const blocks = state.blocks;
+
+  for (let i = blocks.length - 1; i >= 0; --i) {
+    const block = blocks[i];
+
+    if (!block.color || block.garbage) {
+      continue;
+    }
+
+    // Floating and falling:
+    // A block can float or fall if
+    // it is not at ground level and
+    // it is not flashing
+    // and the block (or air) bellow is not swapping and
+    // either beneath there's air or a floating/falling block.
+
+    // Swapping blocks start floating, but wait until
+    // the swap completes before continuing with the fall.
+    const bellow = blocks[i + state.width];
+    if (
+      bellow &&
+      (block.flashTimer < 0) &&
+      !bellow.swapTimer &&
+      (!bellow.color || bellow.floatTimer >= 0)
+    ) {
+      if (block.floatTimer < 0) {
+        if (bellow.color && !block.swapTimer) {
+          block.floatTimer = bellow.floatTimer;
+        } else {
+          block.floatTimer = state.floatTime;
+        }
+        block.chaining = (block.chaining || bellow.chaining);
+      } else if (!block.floatTimer) {
+        // Inherit the timer when falling onto a floating block.
+        // Only fall through air.
+        if (bellow.color) {
+          block.floatTimer = bellow.floatTimer;
+        } else {
+          blocks[i] = bellow;
+          blocks[i + state.width] = block;
+        }
+      }
+    } else {
+      block.floatTimer = -1;
+    }
+  }
+}
+
+function handleChaining(state) {
+  floodFill(state, (block, neighbour) => {
+    if (block.chaining || !block.matching) {
+      return false;
+    }
+    if (neighbour.chaining) {
+      block.chaining = true;
+      return true;
+    };
+  });
+}
+
+function handleTimers(state) {
+  const blocks = state.blocks;
+  let chainMatchMade = false;
+  let chainAlive = false;
+
+  blocks.forEach((block, i) => {
+    const above = blocks[i - state.width];
+    const bellow = blocks[i + state.width];
+
+    if (!block.color) {
+      block.chaining = false;
+    }
+
+    // Chaining ends when a block
+    // lands and
+    // is not flashing and
+    // is not matching and
+    // is not on top of a swapping block.
+    if (
+      block.floatTimer < 0 &&
+      block.flashTimer < 0 &&
+      !block.matching &&
+      (!bellow || !bellow.swapTimer)
+    ) {
+      block.chaining = false;
+    }
+
+    chainAlive = (chainAlive || block.chaining);
+
+    if (block.floatTimer > 0) {
+      --block.floatTimer;
+    }
+
+    if (!--block.flashTimer) {
+      clearBlock(block);
+      if (above && above.color) {
+        above.chaining = true;
+      }
+    }
+    if (block.matching) {
+      block.flashTimer = state.flashTime;
+      if (block.chaining) {
+        chainMatchMade = true;
+      }
+    }
+  });
+
+  clearMatches(blocks);
+
+  if (!chainAlive) {
+    state.chainNumber = 0;
+  }
+  if (chainMatchMade) {
+    state.chainNumber++;
+  }
+}
+
+module.exports = {
+  handleSwapping,
+  findMatches,
+  clearMatches,
+  addRow,
+  handleEvents,
+  handleGravity,
+  handleChaining,
+  handleTimers,
+};
