+++ lib/engine-util.js
@@ -33,3 +33,54 @@ module.exports.printColors = ((game) => {
     }
   });
 });
+
+module.exports.newBlock = (() => {
+  return {
+    color: null,
+    flashTimer: -1,
+    floatTimer: -1,
+    swapTimer: 0,
+    chaining: false,
+    garbage: false,
+  };
+});
+
+module.exports.clearBlock = ((block) => {
+  block.color = null;
+  block.flashTimer = -1;
+  block.floatTimer = -1;
+  block.swapTimer = 0;
+  block.chaining = false;
+  block.garbage = false;
+});
+
+// Iterate over the blocks and their neighbours while the callback returns true
+module.exports.floodFill = ((state, callback) => {
+  const blocks = state.blocks;
+  let floodFillActive = true;
+  while (floodFillActive) {
+    floodFillActive = false;
+    blocks.forEach((block, i) => {
+      const bellow = blocks[i + state.width];
+      const above = blocks[i - state.width];
+      let left;
+      let right;
+
+      if (i % state.width > 0) {
+        left = blocks[i - 1];
+      }
+      if (i % state.width < state.width - 1) {
+        right = blocks[i + 1];
+      }
+
+      [left, right, above, bellow].forEach((neighbour) => {
+        if (!neighbour) {
+          return;
+        }
+        if (callback(block, neighbour)) {
+          floodFillActive = true;
+        }
+      });
+    });
+  }
+});
